PlotQuestion

+++ algebra742.py
@@ -753,7 +753,7 @@ def Assignment(lti=lti, assignment=None,q=None,i=None):
             input_coordinate_pairs = re.split("\)\s*,\s*\(",form.set_of_coordinate_pairs.data)
             input_set_of_coordinate_pairs = set()
             for input_coordinate_pair_string in input_coordinate_pairs:
-                input_coordinate_pair = tuple(int(x.strip("{()} ")) for x in input_coordinate_pair_string.split(","))
+                input_coordinate_pair = tuple(float(x.strip("{()} ")) for x in input_coordinate_pair_string.split(","))
                 input_set_of_coordinate_pairs.add(input_coordinate_pair)
             app.logger.error(input_set_of_coordinate_pairs)
             correct = input_set_of_coordinate_pairs == Parameters['set_of_coordinate_pairs']
